+++ collie/collie.h
@@ -53,7 +53,7 @@ extern int sdport;
 extern int highlight;
 extern int raw_output;
 
-extern uint64_t node_list_version;
+extern uint32_t node_list_version;
 extern struct sd_node node_list_entries[SD_MAX_NODES];
 extern struct sd_vnode vnode_list_entries[SD_MAX_VNODES];
 extern int nr_nodes, nr_vnodes;
